Feat: 운영게시판 API 구현

+++ backend/postings/views.py
@@ -3,55 +3,65 @@ import json
 from django.views  import View
 from django.http   import JsonResponse
 
-from postings.models import FreeBoardPosting, FreeComment, FreeView
+from postings.models import OperatingBoardPosting, OperatingComment, OperatingView
 
-class FreeBoardListView(View):
+class OperatingListView(View):
     """
     * @code writer 조현우
-    * @GET ("/postings/freeboards")
+    * @GET ("/postings/operatings")
     *
     * @returns json
     """
     def get(self, request):
-        postings = FreeBoardPosting.objects.all()
+        user = request.user
+
+        if user.level != 2:
+            return JsonResponse({'message' : 'NO_AUTHENTIFICATION'}, status = 403)
+
+        postings = OperatingBoardPosting.objects.all()
 
         results = [{
             'id'      : posting.id,
             'title'   : posting.title,
             'context' : posting.context[:15],
-            'views'   : FreeView.objects.filter(free_board_posting=posting).count()
+            'views'   : OperatingView.objects.filter(operating_board_posting=posting).count()
         } for posting in postings ]
 
         return JsonResponse({'results' : results}, status = 200)
 
 
-class FreeBoardDetailView(View):
+class OperatingDetailView(View):
     """
     * @code writer 조현우
-    * @GET ("/postings/freeboards/detail")
+    * @GET ("/postings/operatings/detail")
     *
     * @returns json
     """
     def get(self, request, posting_id):
         try:
-            posting = FreeBoardPosting.objects.get(id=posting_id)
-            comments = FreeComment.objects.filter(free_board_posting=posting)
+            user = request.user
+
+            if user.level != 2:
+                return JsonResponse({'message' : 'NO_AUTHENTIFICATION'}, status = 403)
+            
+            posting = OperatingBoardPosting.objects.get(id=posting_id)
+            comments = OperatingComment.objects.filter(operating_board_posting=posting)
 
             result = {
                 'id'       : posting.id,
                 'title'    : posting.title,
                 'context'  : posting.context,
-                'views'    : FreeView.objects.filter(free_board_posting=posting).count(),
+                'views'    : OperatingView.objects.filter(operating_board_posting=posting).count(),
                 'comments' : [comment for comment in comments]
             }
 
             return JsonResponse({'result' : result}, status = 200)
-        except FreeBoardPosting.DoesNotExist:
+        except OperatingBoardPosting.DoesNotExist:
             return JsonResponse({'message' : 'POSTING_DOES_NOT_EXIST'}, status = 400)
 
     """
     * @code writer 조현우
-    * @POST ("/postings/freeboards/detail/<int:posting_id>")
+    * @POST ("/postings/operatings/detail/<int:posting_id>")
     *
     * @returns json
     """
@@ -59,7 +69,10 @@ class FreeBoardDetailView(View):
         try:
             data    = json.loads(request.body)
 
-            posting, is_created = FreeBoardPosting.objects.update_or_create(
+            if user.level != 2:
+                return JsonResponse({'message' : 'NO_AUTHENTIFICATION'}, status = 403)
+
+            posting, is_created = OperatingBoardPosting.objects.update_or_create(
                 title    = data['title'],
                 context  = data['context'],
                 defaults = {
@@ -76,18 +89,23 @@ class FreeBoardDetailView(View):
 
     """
     * @code writer 조현우
-    * @DELETE ("/postings/freeboards/detail")
+    * @DELETE ("/postings/operatings/detail")
     *
     * @returns json
     """
     def delete(self, request):
         try:
-            data    = json.loads(request)
-            posting = FreeBoardPosting.objects.get(id=data["posting_id"])
+            data = json.loads(request)
+            user = request.user
+
+            if user.level != 2:
+                return JsonResponse({'message' : 'NO_AUTHENTIFICATION'}, status = 403)
+
+            posting = OperatingBoardPosting.objects.get(id=data["posting_id"])
             posting.delete()
 
             return JsonResponse({'message' : 'DATA_DELETED'}, status = 204)
         except KeyError:
             return JsonResponse({'message' : 'KEY_ERROR'}, status = 400)
-        except FreeBoardPosting.DoesNotExist:
-            return JsonResponse({'message' : 'DATA_DOES_NOT_EXIST'}, status = 400)
+        except OperatingBoardPosting.DoesNotExist:
+            return JsonResponse({'message' : 'POSTING_DOES_NOT_EXIST'}, status = 400)
